requirements.txt
aiomysql==0.3.2
aiosmtplib==5.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.124.4
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastapi-mail==1.6.1
fastar==0.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
passlib==1.7.4
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyMySQL==1.1.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.11.3
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.47.0
shellingham==1.5.4
six==1.17.0
starlette==0.50.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
